requirements.txt
# List of python dependencies 
slack-bolt
python-dotenv